tests/SDK/SDK.int.2/tng-sdk-validation/requirements.txt
setuptools
pytest
pyaml
coloredlogs
networkx
validators
jsonschema
flask_restplus
flask_cors
flask_cache
requests
